+++ ghost/core/test/unit/server/lib/bootstrap-socket.test.js
@@ -1,6 +1,6 @@
 require('should');
 
-const bootstrapSocket = require('../lib/bootstrap-socket');
+const bootstrapSocket = require('../../../../core/server/lib/bootstrap-socket');
 
 describe('Connect and send', function () {
     it('Resolves a promise for a bad call', function () {
